# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Neural Navigator API
  description: |
    REST API for Neural Navigator - An AI agent prediction game where users compete by
    predicting reinforcement learning agent performance in MiniGrid navigation tasks.

    ## Authentication
    This API uses session-based authentication with secure cookies. After logging in,
    the session cookie is automatically included in subsequent requests.

    ## How the Game Works
    1. Choose an AI agent (DDQN or D3QN)
    2. Predict the number of steps it will take (1-120) or predict it will fail
    3. Watch the AI navigate a 6-room MiniGrid maze
    4. Earn points based on prediction accuracy

    ## Scoring
    - Perfect (0 steps off): 100 points
    - Close (1-10 steps off): 50 points
    - Far (11-20 steps off): 25 points
    - Way Off (21+ steps off): 0 points
    - Correct failure prediction: 50 points
  version: 1.0.0
  contact:
    name: Neural Navigator
    url: https://github.com/TheMichael/MiniGridGame
servers:
  - url: https://neural-navigator.onrender.com
    description: Production server
  - url: http://localhost:5000
    description: Local development server
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: Session-based authentication using secure cookies
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: player1
        email:
          type: string
          format: email
          example: player1@example.com
        total_score:
          type: integer
          example: 500
        games_played:
          type: integer
          example: 10
        best_score:
          type: integer
          example: 100
        is_admin:
          type: boolean
          example: false
        is_active:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
        last_login:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: An error occurred
...
paths:
  /api/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Create a new user account. The first registered user automatically becomes an admin.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - email
                - password
              properties:
                username:
                  type: string
                  example: player1
                  minLength: 3
                  maxLength: 50
                email:
                  type: string
                  format: email
                  example: player1@example.com
                password:
                  type: string
                  format: password
                  minLength: 6
                  example: securepassword123
                skip_auto_login:
                  type: boolean
                  default: false
                  example: false
                  description: Set to true to skip automatic login (useful for admin testing in Swagger UI)
      responses:
        '200':
          description: Registration successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate a user and create a session cookie.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  example: player1
                password:
                  type: string
                  format: password
                  example: securepassword123
                skip_auto_login:
                  type: boolean
                  default: false
                  example: false
                  description: Set to true to skip automatic login (useful for admin testing in Swagger UI)
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Missing credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid credentials or inactive account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/logout:
    post:
      tags:
        - Authentication
      summary: User logout
      description: Clear the user's session cookie and log them out.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                skip_auto_login:
                  type: boolean
                  default: false
                  example: false
                  description: Set to true to skip actual logout (useful for admin testing in Swagger UI)
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
  /api/me:
    get:
      tags:
        - Authentication
      summary: Get current user
      description: Retrieve information about the currently authenticated user.
      security:
        - cookieAuth: []
      responses:
        '200':
          description: User information retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      username:
                        type: string
                        example: player1
                      total_score:
                        type: integer
                        example: 500
                      games_played:
                        type: integer
                        example: 10
                      best_score:
                        type: integer
                        example: 100
                      created_at:
                        type: string
                        format: date-time
                      is_admin:
                        type: boolean
                        example: false
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/forgot-password:
    post:
      tags:
        - Authentication
      summary: Request password reset
      description: Request a password reset link to be sent via email. For security, always returns success even if the user doesn't exist.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - identifier
              properties:
                identifier:
                  type: string
                  example: player1@example.com
                  description: User's email address or username
      responses:
        '200':
          description: Password reset request processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: If the account exists, a reset link has been sent to your email.
        '400':
          description: Missing identifier
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/reset-password:
    post:
      tags:
        - Authentication
      summary: Reset password
      description: Reset a user's password using a valid reset token received via email.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
                - password
              properties:
                token:
                  type: string
                  example: abc123def456ghi789
                  description: Password reset token from email
                password:
                  type: string
                  format: password
                  minLength: 6
                  example: newsecurepassword123
                  description: New password (minimum 6 characters)
      responses:
        '200':
          description: Password reset successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Password reset successfully!
        '400':
          description: Invalid input or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...